requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
asyncpg==0.30.0
beautifulsoup4==4.12.3
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.121.3
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gotrue==2.12.4
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
lxml==5.2.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pdfminer-six==20251107
pdfplumber==0.11.8
pillow==12.0.0
postgrest==0.18.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pyjwt==2.10.1
pyparsing==3.2.5
pypdfium2==5.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
realtime==2.24.0
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.50.0
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tavily-python==0.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.0
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.121.3
fastavro==1.12.1
ffmpeg-python==0.2.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gotrue==2.12.4
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==1.2.3
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.2
langchain-text-splitters==1.1.0
langsmith==0.5.0
mmh3==4.1.0
multidict==6.7.0
nltk==3.9.2
numpy==2.3.5
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pdfminer-six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pinecone==8.0.0
pinecone-client==6.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
pinecone-text==0.11.0
postgrest==0.18.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pyjwt==2.10.1
pyparsing==3.2.5
pypdfium2==5.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
realtime==2.24.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
starlette==0.50.0
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tavily-python==0.5.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid-utils==0.12.0
uvicorn==0.38.0
voyageai==0.3.7
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.0
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.121.3
fastavro==1.12.1
ffmpeg-python==0.2.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gotrue==2.12.4
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==1.2.3
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.2
langchain-text-splitters==1.1.0
langsmith==0.5.0
lxml==6.0.2
mmh3==4.1.0
multidict==6.7.0
nltk==3.9.2
numpy==2.3.5
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pdfminer-six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pinecone==8.0.0
pinecone-client==6.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
pinecone-text==0.11.0
postgrest==0.18.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pyjwt==2.10.1
pyparsing==3.2.5
pypdfium2==5.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
realtime==2.24.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
starlette==0.50.0
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tavily-python==0.5.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid-utils==0.12.0
uvicorn==0.38.0
voyageai==0.3.7
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
